Moved AdditionalPosesPublisher and FramePorts to separate files.
