+++ src/devices/openxrheadset/CMakeLists.txt
@@ -48,6 +48,8 @@ set(yarp_openxrheadset_driver_SRCS
     OpenXrYarpUtilities.cpp
     SingleEyePort.cpp
     EyesManager.cpp
+    AdditionalPosesPublisher.cpp
+    FramePorts.cpp
 )
 
 set(yarp_openxrheadset_driver_HDRS
@@ -55,6 +57,8 @@ set(yarp_openxrheadset_driver_HDRS
     OpenXrYarpUtilities.h
     SingleEyePort.h
     EyesManager.h
+    AdditionalPosesPublisher.h
+    FramePorts.h
 )
 
 set (THRIFTS thrifts/OpenXrHeadsetCommands.thrift)
